Test case for NH-3413 - Clearing a list used by Contains causes subsequent queries to fail
The error is introduced by the call to names.Clear(), this has no effect
on the query, as the query is evaluated eagerly.

It does however break the second query that is performed in a different
session. This error continues to persist until the application is
restarted or the HQL query plan expires from the cache.